+++ lib/core/macros.nim
@@ -492,6 +492,25 @@ proc bindSym*(ident: string | NimNode, rule: BindSymRule = brClosed): NimNode {.
   ## returned even if the symbol is not ambiguous.
   ##
   ## See the `manual <manual.html#macros-bindsym>`_ for more details.
+  ## 
+  ## See also `dynamicBindSym`_.
+
+proc dynamicBindSym*(ident: string | NimNode, rule: BindSymRule = brClosed): NimNode {.
+              magic: "NBindSym", noSideEffect.}
+  ## Creates a node that binds `ident` to a symbol node. The bound symbol
+  ## may be an overloaded symbol.
+  ## if `ident` is a NimNode, it must have `nnkIdent` kind.
+  ## If `rule == brClosed` either an `nnkClosedSymChoice` tree is
+  ## returned or `nnkSym` if the symbol is not ambiguous.
+  ## If `rule == brOpen` either an `nnkOpenSymChoice` tree is
+  ## returned or `nnkSym` if the symbol is not ambiguous.
+  ## If `rule == brForceOpen` always an `nnkOpenSymChoice` tree is
+  ## returned even if the symbol is not ambiguous.
+  ## 
+  ## In contrast to `bindSym`_, `ident` does not have to be a constant value,
+  ## and instead may be a computed value. Note that this behavior is experimental.
+  ##
+  ## See the `manual <manual.html#macros-bindsym>`_ for more details.
 
 proc genSym*(kind: NimSymKind = nskLet; ident = ""): NimNode {.
   magic: "NGenSym", noSideEffect.}